+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for OpenSearch database type
+- Add `database index list` and `database index` commands for managing OpenSearch database indices 
 
 ## [2.7.1] - 2023-05-16
 ### Fixed